+++ src/ConnectKafkaFake.h
@@ -24,7 +24,7 @@ public:
 
   int getNumberOfTopicPartitions(std::string TopicName) override;
 
-  void subscribeAtOffset(int64_t Offset, std::string TopicName) override;
+  void subscribeAtOffset(int32_t Offset, std::string TopicName) override;
 
   void subscribeToLastNMessages(int64_t NMessages, const std::string &TopicName,
                                 int Partition) override;
